% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{apply_binning}
\alias{apply_binning}
\title{Apply binning}
\usage{
apply_binning(bin, x)
}
\arguments{
\item{bin}{A \code{binning} object}

\item{x}{A variable to bin.}
}
\description{
Apply binning
}
