% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning.character}
\alias{binning.character}
\title{Grouping or binning character/factors variables}
\usage{
\method{binning}{character}(x, y, min.p = 0.2, alpha = 0.05,
  na_level = "(Missing)", ...)
}
\arguments{
\item{x}{A variable to binning.}

\item{y}{A binary variable same length as \code{score}.}

\item{min.p}{Minimal proportion in a group (a ctree_control argument).}

\item{alpha}{Minimal critetion (a ctree_control argument).}

\item{na_level}{String for determinar de NA level.}

\item{...}{Extra arguments to \code{ctree_control}.}
}
\description{
Grouping or binning character/factors variables
}
