% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning.numeric}
\alias{binning.numeric}
\title{Grouping or binning for numeric variables}
\usage{
\method{binning}{numeric}(x, y, min.p = 0.1, alpha = 0.05,
  na_level = "(Missing)", ...)
}
\arguments{
\item{x}{A variable to binning.}

\item{y}{A binary variable same length as \code{score}.}

\item{min.p}{Minimal proportion in a group (a ctree_control argument).}

\item{alpha}{Minimal critetion (a ctree_control argument).}

\item{na_level}{String for determinar de NA level.}

\item{...}{Extra arguments to \code{ctree_control}.}
}
\description{
Grouping or binning for numeric variables
}
\examples{

data(german_credit)
binning(german_credit$duration_in_month, german_credit$good_bad)
binning(german_credit$credit_amount, german_credit$good_bad)

}
