% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_table.R
\name{bivariate_table}
\alias{bivariate_table}
\title{Calculate bivariate table from a vector and a response variable}
\usage{
bivariate_table(x, target, add.prop.test = TRUE, conf.level = 0.95)
}
\arguments{
\item{x}{A variable.}

\item{target}{A binary variable same length as \code{x}.}

\item{add.prop.test}{Boolean to add or not the \code{prop.test} by
categories.}

\item{conf.level}{Confidence level of the returned confidence interval.}
}
\description{
Calculate bivariate table from a vector and a response variable
}
\examples{

data(german_credit)

bivariate_table(german_credit$purpose, german_credit$good_bad)


x <- cut(german_credit$duration_in_month, c(0, 6, 12, 24, 36, Inf))
target <- german_credit$good_bad

bivariate_table(x, target)

}
