% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe}
\usage{
describe(data)
}
\arguments{
\item{data}{A data frame to describe.}
}
\description{
Describe
}
\examples{

data("german_credit")

describe(german_credit)

describe(german_credit[c("good_bad", "purpose")])

describe(german_credit[c("good_bad", "duration_in_month")])

describe(german_credit[c("status_of_existing_checking_account", "savings_account_or_bonds")])

\dontrun{
data(chileancredit, package = "smbinning")

d <- subset(chileancredit, select = -c(period))

describe(d)
}

}
