% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_bivariate}
\alias{describe_bivariate}
\title{Describe Bivariate}
\usage{
describe_bivariate(data, target, p.min = 0.05, alpha = 0.05,
  verbose = TRUE)
}
\arguments{
\item{data}{A data frame to describe.}

\item{target}{The target to describe.}

\item{p.min}{Minimal proportion in a group (a ctree_control argument).}

\item{alpha}{Minimal critetion (a ctree_control argument).}

\item{verbose}{verbose.}
}
\description{
Describe Bivariate
}
\examples{

data("german_credit")

dbiv <- describe_bivariate(german_credit, target = good_bad)
dplyr::glimpse(dbiv)

describe_bivariate(german_credit, target = good_bad, alpha = 0.5)


\dontrun{

data(chileancredit, package = "smbinning")
data <- subset(chileancredit, select = -c(period))
describe_bivariate(data, target = fgood)

}

}
