% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_selection_raw}
\alias{feature_selection_raw}
\title{Perform a very raw Feature Selection}
\usage{
feature_selection_raw(data, target = TARGET, max_concentration_limit = 0.9,
  min_iv = 0.2, verbose = TRUE, ...)
}
\arguments{
\item{data}{A data frame to perform the feature selection.}

\item{target}{The target.}

\item{max_concentration_limit}{A maximum concentration limit.}

\item{min_iv}{Minimum Information Value to accpet}

\item{verbose}{verbose.}

\item{...}{Additional parameters for \code{binning} function.}
}
\description{
Perform a very raw Feature Selection
}
\examples{

data("german_credit")

fs <- feature_selection(german_credit, good_bad, max_concentration_limit = 0.4, min_iv = 0.1)

table(fs$decision)

}
