% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irks-package.R
\name{irks-exports}
\alias{irks-exports}
\alias{\%>\%}
\title{irks exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the irks package to avoid listing the magrittr
as Depends of irks.
}
