% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irks-package.R
\docType{package}
\name{irks}
\alias{irks}
\alias{irks-package}
\title{Functions To Develop Credit Risk Models}
\description{
Functions To Develop Credit Risk Models
}
\author{
Joshua Kunst (@jbkunst)
}
