% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_plot}
\alias{model_plot}
\title{Plot a model}
\usage{
model_plot(model)
}
\arguments{
\item{model}{A glm logistic model}
}
\description{
Plot a model
}
\examples{

data("german_credit")

model <- glm(
  good_bad ~ purpose + present_employment_since + credit_history,
  data = german_credit, family = binomial
  )

model_plot(model)

}
