% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_scorecad}
\alias{model_scorecad}
\title{Creating a scorecard from logistic model}
\usage{
model_scorecad(model, pdo = 20, score0 = 600, pdo0 = 20)
}
\arguments{
\item{model}{A glm logistic model}

\item{pdo}{default 20}

\item{score0}{default 600}

\item{pdo0}{default to 20}
}
\description{
Creating a scorecard from logistic model
}
