% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_scorecard.R
\name{model_scorecard}
\alias{model_scorecard}
\title{Creating a scorecard from logistic model}
\usage{
model_scorecard(model, pdo = 20, score0 = 600, pdo0 = 50/1,
  turn.orientation = FALSE)
}
\arguments{
\item{model}{A glm logistic model}

\item{pdo}{default 20}

\item{score0}{default 600}

\item{pdo0}{default to 50/1}

\item{turn.orientation}{change the orientation of the scorecard points}
}
\description{
The defaults are given by de Siddiqi example.
}
\examples{

data("german_credit")

model <- glm(
  good_bad ~ purpose + present_employment_since + credit_history,
  data = german_credit, family = binomial
  )

model_scorecard(model)

}
