% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{print.binning}
\alias{print.binning}
\title{Generic method to print a binnig object}
\usage{
\method{print}{binning}(x, ...)
}
\arguments{
\item{x}{A binning object.}

\item{...}{Extra arguments.}
}
\description{
Generic method to print a binnig object
}
\examples{

data(german_credit)
bin <- binning(german_credit$duration_in_month, german_credit$good_bad)
bin

x <- runif(500)
y <- rbinom(500, 1, x)
bin <- binning(x, y, alpha = .1)
bin

}
