% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-auxiliars.R
\name{roc_data}
\alias{roc_data}
\title{Function to get data to plot a roc curve}
\usage{
roc_data(score, target)
}
\arguments{
\item{score}{A numeric vector of predictions.}

\item{target}{A binary variable same length as \code{score}.}
}
\description{
Function to get data to plot a roc curve
}
