% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{scorecard}
\alias{scorecard}
\title{Creating a scorecard from logistic model}
\usage{
scorecard(model, pdo = 50:1, score0 = 600, pdo0 = 20)
}
\arguments{
\item{model}{A glm logistic model}

\item{pdo}{default 20}

\item{score0}{default 600}

\item{pdo0}{default to 20}
}
\description{
The defaults are given by de Siddiqi example.
}
