% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning.character}
\alias{binning.character}
\title{Grouping or binning character/factors variables}
\usage{
\method{binning}{character}(x, y, p.min = 0.05, alpha = 0.05,
  na_level = "(Missing)", keeptree = FALSE, ...)
}
\arguments{
\item{x}{A variable to binning.}

\item{y}{A binary variable same length as \code{score}.}

\item{p.min}{Minimal proportion in a group (a ctree_control argument).}

\item{alpha}{Minimal critetion (a ctree_control argument).}

\item{na_level}{String for determinar de NA level.}

\item{keeptree}{logical to keep tree.}

\item{...}{Extra arguments to \code{ctree_control}.

data(german_credit)
binning(german_credit$purpose, german_credit$good_bad)}
}
\description{
Grouping or binning character/factors variables
}
