% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{Supervised binning}
\usage{
binning(x, ...)
}
\arguments{
\item{x}{A variable to binning.}

\item{...}{Additional arguments for the binning process.}
}
\description{
Automatic supervised binning
}
