% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_table.R
\name{bivariate_table}
\alias{bivariate_table}
\title{Calculate bivariate table from a vector and a response variable}
\usage{
bivariate_table(x, target)
}
\arguments{
\item{x}{A variable.}

\item{target}{A binary variable same length as \code{x}.}
}
\description{
Calculate bivariate table from a vector and a response variable
}
