% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_summary}
\alias{model_summary}
\title{Get a summary of the logistic glm model considering reference levels}
\usage{
model_summary(model)
}
\arguments{
\item{model}{A glm logistic model}
}
\description{
Get a summary of the logistic glm model considering reference levels
}
\examples{

data("german_credit")

model <- glm(
  good_bad ~ purpose + present_employment_since + credit_history,
  data = german_credit, family = binomial
  )

model_summary(model)

}
