% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{plot.binning}
\alias{plot.binning}
\title{Generic method to plot a binnig object}
\usage{
\method{plot}{binning}(x, ...)
}
\arguments{
\item{x}{A binning object.}

\item{...}{Extra arguments.}
}
\description{
Generic method to plot a binnig object
}
\examples{

data(german_credit)
bn <- binning(german_credit$duration_in_month, german_credit$good_bad)

plot(bn)

}
